+++ assets/javascripts/kitten/karl/intersection-observer/stories.js
@@ -22,7 +22,7 @@ const ObservedComponent = ({ children, ...props }) => {
   )
 }
 
-storiesOf('IntersectionObserverClass', module).add(
+storiesOf('IntersectionObserver/IntersectionObserverClass', module).add(
   'using IntersectionObserverClass hook',
   () => {
     return (