+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `useContinuousIntersectionObserver` hook.
+- Feature: Add `ContinuousIntersectionObserver` component with Render Prop.
+
 ## [2.51.0] - 2020-01-20
 
 Features: